split plotting logically
