+++ qcodes/plotting/__init__.py
@@ -5,12 +5,13 @@ For the current dataset see :class:`qcodes.dataset.plot_dataset`
 For the legacy dataset see :class:`qcodes.plots`
 """
 
+from .auto_range import auto_range_iqr
+from .axis_labels import find_scale_and_prefix
 from .matplotlib_helpers import (
     apply_auto_color_scale,
     apply_color_scale_limits,
     auto_color_scale_from_config,
 )
-from .plotting import auto_range_iqr, find_scale_and_prefix
 
 __all__ = [
     "apply_auto_color_scale",
